add option of whether to destroy infrastructure

+++ Jenkinsfile
@@ -1,22 +1,20 @@
 pipeline {
+
     agent {
     label 'eng110-jenkins-worker'
     }
 
     parameters {
-        booleanParam(name: 'autoApprove', defaultValue: false, description: 'Automatically run apply after generating plan?')
-        booleanParam(name: 'destroy', defaultValue: false, description: 'Destroy Terraform build?')
+      booleanParam defaultValue: true, description: 'Carry out terraform destroy?', name: 'Destroy'
     }
 
     stages {
-        stage('checkout') {
+        stage('Checkout') {
             steps {
                  script{
                         dir("terraform")
                         {
                             checkout([$class: 'GitSCM', branches: [[name: '*/main']], extensions: [], userRemoteConfigs: [[url: 'https://github.com/samuel-walters/Complete-CICD']]])
-                            sh 'echo $(pwd)'
-                            sh 'echo $(ls)'
                         }
                     }
                 }
@@ -26,45 +24,43 @@ pipeline {
             steps {
                 script{
                     withCredentials([aws(accessKeyVariable: 'AWS_ACCESS_KEY_ID', credentialsId: 'my-aws-credentials', secretKeyVariable: 'AWS_SECRET_ACCESS_KEY')]) {
-                    dir("terraform")
-                        {
-                                sh 'echo HELLOOOOO'
-                                sh 'echo $(ls)'
-                                sh 'terraform init -input=false'
-                                sh "terraform plan -input=false -out tfplan "
-                                sh 'terraform show -no-color tfplan > tfplan.txt'
+                        dir("terraform") {
+                            sh 'terraform init -input=false'
+                            sh "terraform plan -input=false -out tfplan "
+                            sh 'terraform show -no-color tfplan > tfplan.txt'
                         }
-                        }
+                    }
                 }
-                }
-            }
+            }  
+        }
         
 
         stage('Apply') {
-            
             steps {
                 script {
                     withCredentials([aws(accessKeyVariable: 'AWS_ACCESS_KEY_ID', credentialsId: 'my-aws-credentials', secretKeyVariable: 'AWS_SECRET_ACCESS_KEY')]) {
-                    dir("terraform") {
+                        dir("terraform") {
                             sh "terraform apply -input=false tfplan"
-                    }
+                        }
                     }
                 }
             }
         }
         
         stage('Destroy') {
-        
-        steps {
-            script {
-                    withCredentials([aws(accessKeyVariable: 'AWS_ACCESS_KEY_ID', credentialsId: 'my-aws-credentials', secretKeyVariable: 'AWS_SECRET_ACCESS_KEY')]) {
-                    dir("terraform") {
-                           sh "terraform destroy --auto-approve"
+            steps {
+                script {
+                    if(params.Destroy == true){  
+                        withCredentials([aws(accessKeyVariable: 'AWS_ACCESS_KEY_ID', credentialsId: 'my-aws-credentials', secretKeyVariable: 'AWS_SECRET_ACCESS_KEY')]) {
+                            dir("terraform") {
+                            sh "terraform destroy --auto-approve"
+                            }
+                        }
+                    } else {
+                        echo "Skipping terraform destroy"
                     }
-                    }
+                } 
             }
         }
     }
-
-  }
 }
